claudecode/requirements.txt
# Requirements for claudecode v2.0 - Claude Code Security Audit Tool

# -- CORE --
PyGithub==2.1.1
requests==2.31.0
anthropic==0.39.0
openai==1.12.0
pydantic==2.5.3
tenacity==8.2.3
aiohttp==3.9.1
PyYAML==6.0.1
packaging==23.2

# -- WEB INTERFACE --
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6

# -- SECURITY UTILS --
detect-secrets==1.4.0
sarif-om==1.0.4
typing-extensions==4.9.0